+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.2.4] - 2026-02-08
+### ⚡ Changed
+- **Port**: use dynamic port for backend in release mode
+- **Desktop App**: use pyinstaller for faster build
+
+### 🐛 Fixed
+- **Social Messaging**: fix social messaging setting's view crashes
+
+
 ## [v0.2.3] - 2026-02-03
 
 ### 🚀 Added